+++ include/thingset++/ThingSetProperty.hpp
@@ -113,12 +113,12 @@ public:
         return _name;
     }
 
-    unsigned getId() const override
+    uint16_t getId() const override
     {
         return _id;
     }
 
-    unsigned getParentId() const override
+    uint16_t getParentId() const override
     {
         return _parentId;
     }
@@ -185,12 +185,12 @@ public:
         return _name;
     }
 
-    unsigned getId() const override
+    uint16_t getId() const override
     {
         return _id;
     }
 
-    unsigned getParentId() const override
+    uint16_t getParentId() const override
     {
         return _parentId;
     }
@@ -282,12 +282,12 @@ public:
         return _name;
     }
 
-    unsigned getId() const override
+    uint16_t getId() const override
     {
         return _id;
     }
 
-    unsigned getParentId() const override
+    uint16_t getParentId() const override
     {
         return _parentId;
     }
